+++ platform/src/EducationPlatform.js
