+++ platform/src/index.js
@@ -1,6 +1,6 @@
 
 /*global TOKEN_SERVER_URL -- is set by environment variable*/
-import { EducationPlatform } from "./Playground";
+import { EducationPlatform } from "./EducationPlatform.js";
 
 const TOKEN_HANDLER_URL = TOKEN_SERVER_URL || "http://127.0.0.1:10000";
 let urlParameters = new URLSearchParams(window.location.search);
